requirements.txt
fastapi
uvicorn
supabase
google-generativeai
python-dotenv
pydantic-settings
google-cloud-speech
google-cloud-texttospeech
python-multipart